College Visitors Information Portal
Frontend : PYTHON and backend : MYSQL
The College Visitor Portal is a digital platform designed to facilitate and
streamline the process of managing and tracking visitors to a college
campus. using ( Flask ,tkinter.messagebox ).
It provides a convenient and efficient way for visitors to check-in, obtain
visitor passes, and access relevant information during their visit.